### 2020-09-10 #2659 @whedon
[PRE REVIEW]: c212: An R Package for the Detection of Safety Signals in Clinical Trials Using Body-Systems (System Organ Classes)
**Submitting author:** @rcarragh (<a href="http://orcid.org/0000-0002-0120-625X">Raymond Bernard Carragher</a>)
**Repository:** <a href="https://github.com/rcarragh/c212" target ="_blank">https://github.com/rcarragh/c212</a>
**Version:** 0.98
