### 2020-09-10 #2659 @danielskatz
@rcarragh - As you can see by comments in this issue from whedon, there are some problems to be fixed.

In the repo, I see no license.  I also see no README.

In the paper, you need to change the multiple affiliation line to have commas between affiliations - see https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

You also have some extra metadata in the paper which may cause problems once the affiliations part is solved.

Finally, there are 2 DOI issues in the bib file.

Please make changes to the paper (and bib), then add ```@whedon generate pdf``` as a new comment here to regenerate the PDF.  Feel free to iterate.

I will mark this as paused for now.  Once you think you are ready to proceed, please tag @openjournals/joss-eics to let me (or the AEiC on duty) know.
