+++ paper/paper.bib
@@ -7,7 +7,7 @@
    pages = {289-300},
    abstract = {The common approach to the multiplicity problem calls for controlling the familywise error rate (FWER). This approach, though, has faults, and we point out a few. A different approach to problems of multiple significance testing is presented. It calls for controlling the expected proportion of falsely rejected hypotheses-the false discovery rate. This error rate is equivalent to the FWER when all hypotheses are true but is smaller otherwise. Therefore, in problems where the control of the false discovery rate rather than that of the FWER is desired, there is potential for a gain in power. A simple sequential Bonferroni-type procedure is proved to control the false discovery rate for independent test statistics, and a simulation study shows that the gain in power is substantial. The use of the new procedure and the appropriateness of the criterion are illustrated with examples.},
    ISSN = {00359246},
-   DOI = {10.2307/2346101},
+   DOI = {10.1111/j.2517-6161.1995.tb02031.x},
    url = {http://www.jstor.org/stable/2346101},
    year = {1995},
    type = {Journal Article}
@@ -63,7 +63,7 @@ Stat Med. 2012 Aug 15;31(18):1918-30. doi: 10.1002/sim.5310. Epub 2012 Mar 13.},
    number = {2},
    pages = {405-415},
    ISSN = {0378-3758},
-   DOI = {http://dx.doi.org/10.1016/j.jspi.2007.06.006},
+   DOI = {10.1016/j.jspi.2007.06.006},
    url = {http://www.sciencedirect.com/science/article/pii/S0378375807002510},
    year = {2008},
    type = {Journal Article}
