### 2020-09-10 1caee41
Removed cls and output type
