### 2020-09-10 #2659 @rcarragh
@openjournals/joss-eics

Apologies for the errors in the author affiliations section and any additional overhead caused.
(I couldn't see a full compiled version through the whedon web test interface before I submitted the paper, although I could generate a pdf wth pandoc.)

I've made the following changes:

1. I have fixed the affiliations section.
2. I have added a LICENSE file to the repository.
3. I have added a README file to the repository.
4. I have updated paper.bib to address the 2 DOI issues.
5. I have removed metadata from paper.md (csl, output, header-includes)

I have been able to generate a PDF as suggested and to me it looks ok, so hopefully this has addressed the issues reported.
