### 2020-09-11 #2659 @csoneson
👋 @rcarragh - I will handle this submission. Do you have any suggestions for suitable reviewers (e.g., from the spreadsheet linked in the first post above)?

Looking at your repository, it seems that you have different licenses indicated in the `DESCRIPTION` file and in the repository itself - could you please harmonize that? I also notice that your package doesn't seem to have any unit tests. We strongly suggest to have these; at the very least, there must be an objective way for the reviewers to verify the correct functionality of the package (see e.g. the review criteria [here](https://joss.readthedocs.io/en/latest/review_checklist.html)).
