### 2020-09-11 #2659 @rcarragh
@csoneson - I will make these updates (licence and tests) and let you know when they are complete.
Thanks
