+++ test/c212.BB.interim/default_parameters/check_test.sh
@@ -1,0 +1,33 @@
+#!/bin/bash
+
+diff ./test/RAISED_AES ./baseline/RAISED_AES
+if [ $?  -ne 0 ]
+then
+	exit 1
+fi
+
+diff ./test/conv.dat ./baseline/
+if [ $?  -ne 0 ]
+then
+	exit 1
+fi
+
+diff ./test/ptheta80.dat ./baseline/
+if [ $?  -ne 0 ]
+then
+	exit 1
+fi
+
+diff ./test/ptheta.dat ./baseline/
+if [ $?  -ne 0 ]
+then
+	exit 1
+fi
+
+diff ./test/summary.dat ./baseline/
+if [ $?  -ne 0 ]
+then
+	exit 1
+fi
+
+
