+++ test/error_control/run.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+cd ./test
+R -q --no-save <<!
+source("run.r")
+!
