+++ test/c212.BB/initial_values/cleanup.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cd test
+
+rm -f conv.dat  ptheta80.dat  ptheta.dat  summary.dat RAISED_AES

[…]
