### 2020-09-15 98668b4
Fix batch file directories
