### 2020-09-16 013e431
Fix error control check
