### 2020-09-16 #2659 @rcarragh
@csoneson - I have fixed the DESCRIPTION/LICENSE file issue and added a new directory to the repository called "test".
This contains a README explaining the contents of the directory.
The are tests for package functionality, covering all the main use cases and there are demonstration code files for the package in the
subdirectory ./demo.

Starting from the bottom of the spreadsheet a couple of possible reviewers would be:

ChristopherLucas (R/C++/Bayesian statitics)
emilydolson (C++/R)

but someone with a statistics/R/C++ background should be suitable.
