### 2020-09-17 #2659 @csoneson
👋 @emilydolson, @rowlandseymour, @stulacy, @rrrlw - would two of you be interested in reviewing this submission for JOSS?

> c212: An R Package for the Detection of Safety Signals in Clinical Trials Using Body-Systems
