### 2020-09-17 #2659 @emilydolson
I can review it, but I don't have any experience with clinical trial design, so I may not be the best choice (especially if the goal would be for me to focus on the methodology).
