### 2020-09-18 #2659 @csoneson
Thanks @rrrlw and @emilydolson for your availability! Let me try to find a complementary third reviewer who would be more geared towards clinical trial methodology - I'll get back to you.
