### 2020-09-18 #2659 @rowlandseymour
I'm in the same position as Emily. I'm happy to review it, but don't have any experience with clinical trials.
