### 2020-09-28 #2659 @csoneson
Hi again - turns out it's not so easy to find a reviewer with both clinical trial methodology expertise and free time :). I have some leads that are still open, and I will keep looking (we certainly want this aspect evaluated as well), but to avoid holding up the submission at this stage, I suggest that we start the review in the meanwhile - @rrrlw and @emilydolson, are you still available and willing to review the package (if so, @rowlandseymour, I hope we can come back to you for another submission in the future - thanks for volunteering!)
