### 2020-10-03 rcarragh/c212#1 @rrrlw
Missing parts of GPL-3
Please fill in lines 196, 197, 212, and 216 in the LICENSE file.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2706).
