### 2020-10-03 #2706 @rrrlw
N.B. The paper has two authors and the GitHub repo has one contributor. Will trust submitting author's judgment and assume that both authors had substantial contributions and that the author list in the paper is appropriate (checking box in list above).
