### 2020-10-03 rcarragh/c212#2 @rrrlw
Dependency versioning
Adding version requirements to dependency lists in DESCRIPTION can help improve package robustness (in case older dependency versions break code). Would you be willing to add a version requirement for the `coda` dependency? (e.g. `coda >= 0.19` or whichever min version you choose)

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2706).
