### 2020-10-03 rcarragh/c212#3 @rrrlw
R CMD check WARNINGs due to license
Previous DESCRIPTION versions list `GPL (>=2)` as the package license; currently, DESCRIPTION refers to the file `LICENSE`, which (I think) is usually reserved for non-standard licenses (the file itself contains GPL-3, a standard FOSS license).

This results in 2 WARNINGs in `R CMD check --as-cran` for c212 on my local machine with the following text:
```
* checking CRAN incoming feasibility ... WARNING
Change to non-FOSS package license.
New license:
  LICENSE
Old license:
  GPL (>= 2)
* checking DESCRIPTION meta-information ... WARNING
Non-standard license specification:
  LICENSE
Standardizable: FALSE
```

IMO, this could be fixed by switching the package license in DESCRIPTION to `GPL-3`; however, `GPL-3 + file LICENSE` could also be appropriate.

Can use [this chapter from the *R Packages* book](https://r-pkgs.org/license.html) as a reference.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2706).
