### 2020-10-03 rcarragh/c212#4 @rrrlw
Contribution guidelines
One of JOSS's submission requirements is the presence of clear community guidelines for potential contributors. Would you be willing to add appropriate guidelines to this repo? (e.g. in `CONTRIBUTING.md` and/or `CODE_OF_CONDUCT.md` files)

If unsure where to start, I think it would be helpful to check out [Contributor Covenant](https://www.contributor-covenant.org).

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2706).
