+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: c212
 Type: Package
 Title: Methods for Detecting Safety Signals in Clinical Trials Using Body-Systems (System Organ Classes)
-Version: 0.98
-Date: 2020-09-04
+Version: 0.99
+Date: 2020-10-06
 Authors@R:  person(given = "Raymond",
              family = "Carragher",
              role = c("aut", "cre"),
@@ -10,8 +10,8 @@ Authors@R:  person(given = "Raymond",
              comment = c(ORCID = "0000-0002-0120-625X"))
 Author: Raymond Carragher [aut, cre] (<https://orcid.org/0000-0002-0120-625X>)
 Maintainer: Raymond Carragher <rcarragh@gmail.com>
-Depends: coda
+Depends: coda (>= 0.19)
 Encoding: UTF-8
 LazyData: true
 Description: Methods for detecting safety signals in clinical trials using groupings of adverse events by body-system or system organ class. This work was supported by the Engineering and Physical Sciences Research Council (UK) (EPSRC) [award reference 1521741] and Frontier Science (Scotland) Ltd. The package title c212 is in reference to the original Engineering and Physical Sciences Research Council (UK) funded project which was named CASE 2/12.
-License: LICENSE
+License: GPL-3 + file LICENSE
