### 2020-10-06 rcarragh/c212#1 @rcarragh
I have updated the LICENSE file (see also #3). After "END OF TERMS AND CONDITIONS" I've added

Year: 2020
COPYRIGHT HOLDER: Raymond Carragher
Contact: rcarragh@gmail.com
