### 2020-10-06 rcarragh/c212#3 @rcarragh
I can replicate the error.
Some background as to why it happens: when the package was originally released to CRAN in 2017, I specified the licence as GPL (>=2) in the DESCRIPTION file. And it remanied as this up to until I submitted my paper to JOSS. As part of the pre-review process I was asked to include a LICENSE file. I couldn't find a suitable file for GPL >=2 (really multiple licences) so I included the GPL-3 file as the sucessor licence.

I can fix the issue by following your suggestion of:

License: GPL-3 + file LICENSE

in the DESCRIPTION file and I've added

Year: 2020
COPYRIGHT HOLDER: Raymond Carragher
Contact: rcarragh@gmail.com

to the LICENSE file after "END OF TERMS AND CONDITIONS"

This doesn't give me any warnings or notes when when I build the package with "check --as-cran".

Note: this will also affect the "Missing parts of GPL-3 #1" open issue.
