### 2020-10-06 rcarragh/c212#2 @rcarragh
I've updated the DESCRIPTION to code (>= 0.19) as you suggest. Older version should be OK but I agree that this is better.
