+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# Contributing to `c212`
+
+The author is interested in extending the software to include new methods, particularly the area of safety analysis, and would welcome collaborations in this area. 
+
+## Issues and Extensions
+
+For any detected bugs, issues or possible extensions please feel free to create an issue in the c212 github repository or contact the author directly.
+
+When reporting issues, if possible, please include a minimal example with a small data set which reproduces the issue.
+
