### 2020-10-06 rcarragh/c212#4 @rcarragh
I am happy to do this.
I have added CODE_OF_CONDUCT.md and CONTRIBUTING.md to the repository.
