### 2020-10-07 #2706 @rcarragh
I have made some changes to address issues raised above: rcarragh/c212#1, rcarragh/c212#2. rcarragh/c212#3, rcarragh/c212#4.
