### 2020-10-21 rcarragh/c212#3 @rrrlw
You've described an appropriate fix, appreciate the context. Changes look great, thank you!
