+++ DESCRIPTION
@@ -13,5 +13,12 @@ Maintainer: Raymond Carragher <rcarragh@gmail.com>
 Depends: coda (>= 0.19)
 Encoding: UTF-8
 LazyData: true
-Description: Methods for detecting safety signals in clinical trials using groupings of adverse events by body-system or system organ class. This work was supported by the Engineering and Physical Sciences Research Council (UK) (EPSRC) [award reference 1521741] and Frontier Science (Scotland) Ltd. The package title c212 is in reference to the original Engineering and Physical Sciences Research Council (UK) funded project which was named CASE 2/12.
+Description: Provides a self-contained set of methods to aid clinical trial
+    safety investigators, statisticians and researchers, in the early
+    detection of adverse events using groupings by body-system or system
+    organ class. This work was supported by the Engineering and Physical
+    Sciences Research Council (UK) (EPSRC) [award reference 1521741] and
+    Frontier Science (Scotland) Ltd. The package title c212 is in reference
+    to the original Engineering and Physical Sciences Research Council (UK)
+    funded project which was named CASE 2/12.
 License: GPL-3 + file LICENSE
