### 2020-10-25 bec1b87
omit redundant information
