### 2020-10-28 rcarragh/c212#6 @rrrlw
Discuss state of the field
Re: checkbox in JOSS review under section "Software paper" and bullet "State of the field" - are there other software libraries that fulfill a similar function and should be mentioned?

If not, please close this issue with a brief response/explanation.

Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/2706).
