### 2020-10-29 rcarragh/c212#6 @rcarragh
Briefly, I am not aware of a package providing similar set of methods. This was actually part of the rationale for developing the package. Below is a sort of broader description of my understanding of the relationship between c212 and the state of the field.

State of the field:
(Recap: Safety in clinical trials is usually characterised by the incidence (or occurrence) of adverse events (AEs).)

Typically, in a trial’s clinical study report(s) (CSRs), adverse events are just presented in a tabulated form. A number of AEs may possibly have specific hypotheses associated with them and be examined individually, but the majority won’t (leading to the possibility of a missing an AE associated with treatment).

While methods have been developed for analysing AEs in general using body-systems or system organ classes, they have not been routinely being applied to AE data. One issue being that for many of these methods there were no readily available software implementations, leading, if they were to be used at all, to an overhead when trying to apply them on an ad-hoc basis. Part of the project that led to the c212 package was to address this issue and implement some of the more promising methods in a manner which would allow routine application. I am not aware of a similar cohesive collection of methods implemented in software.

I’ve tried to capture this in the last paragraph of the “Statement of Need”, where I try to indicate where c212 lies in the general “State of the Field”. There are a few references to how c212 is used or considered here. The software has been referenced in “Biopharmaceutical Applied Statistics Symposium 2018”, Vol2, Chapter 11, which discusses the methods in the overall context of safety analysis and available tools. c212 has also used as part of recent research projects (Tan et al 2019/2020).
