### 2020-10-29 #2706 @rcarragh
@rrrlw - I've added a comment in rcarragh/c212#6 to address your question about state of the field.
I didn't close the issue so if you want to add anything further of course please do, but if you are happy with the response then we can close it.
