### 2020-11-13 #2706 @MelvinSMunsaka
I will provide my review over the weekend…had some COVID-19 issues as result of getting into contact with cousin who is asymptomatic.

Also, I was taking a closer look at the package, trying out some example and also double checking how we implemented the c212 package here: (https://visual-analytics.shinyapps.io/index/ - see the AE Line Plot tab).

Melvin
