### 2020-11-25 #2706 @MelvinSMunsaka
For some reason, I am not able to add my comments via Github even after I logged into my account.

Here are my few comments:

•	The articles should explain more/correctly about MedDRA and the SOC/PT in the sentence “A number of methods, which use groupings of adverse events by body-system or System Organ Class” as this is key/central to the methods presented, especially some of the Bayesian approaches.

•	The reference “Fries, M., Kracht, K., & Li, J. (2016). Safety monitoring methodology in the premarketing setting. Proceedings of JSM, 2247–2269. Retrieved from https://ww2.amstat.org/MembersOnly/proceedings/2016/data/assets/pdf/389675.pdf” is incorrect. It should be “Fries, M., Kracht, K. K., Li, J., Munsaka, M. S., Sanchez-Kam, M., Singh, K. P., Wang, W., Whalen, E., Zhou. K. (2016). Safety Monitoring Methodology in the Premarketing Setting JSM 2016 Proceedings, Biopharmaceutical Section, 2247 - 2269.  https://ww2.amstat.org/MembersOnly/proceedings/2016/data/assets/pdf/389675.pdf”

•	I stress tested with data sets with a large number of AEs and in some cases the computations crawled to a halt, especially where using the Bayesian method. May want to mention this in the paper since in clinical trial settings the number of AEs can be very large especially in certain therapeutic areas or integrated data settings. As an FYI, we have  implemented the package here: https://visual-analytics.shinyapps.io/index/ (see AE Line Plot tab) as part of the via the ASA Biopharm Safety WG ongoing work.

•	In terms of the checklist above, everything checks out except for the performance issues noted above. I will complete the checklist once I can get it to work. It seems it is greyed out even when I tried changing the “watch” options.
