### 2020-11-25 #2706 @csoneson
Thanks @MelvinSMunsaka! I will re-invite you to this issue - you have to accept the invitation in order to be able to check the boxes. Let me know if it still doesn't work.
