### 2020-11-25 rcarragh/c212#7 @rcarragh
The articles should explain more/correctly about MedDRA and the SOC/PT
The articles should explain more/correctly about MedDRA and the SOC/PT in the sentence “A number of methods, which use groupings of adverse events by body-system or System Organ Class” as this is key/central to the methods presented, especially some of the Bayesian approaches.
