### 2020-11-25 rcarragh/c212#8 @rcarragh
The reference “Fries, M., Kracht, K., & Li, J. (2016). Safety monitoring methodology in the premarketing setting. Proceedings of JSM, 2247–2269. Retrieved from https://ww2.amstat.org/MembersOnly/proceedings/2016/data/assets/pdf/389675.pdf” is incorrect.
The reference “Fries, M., Kracht, K., & Li, J. (2016). Safety monitoring methodology in the premarketing setting. Proceedings of JSM, 2247–2269. Retrieved from https://ww2.amstat.org/MembersOnly/proceedings/2016/data/assets/pdf/389675.pdf” is incorrect. It should be “Fries, M., Kracht, K. K., Li, J., Munsaka, M. S., Sanchez-Kam, M., Singh, K. P., Wang, W., Whalen, E., Zhou. K. (2016). Safety Monitoring Methodology in the Premarketing Setting JSM 2016 Proceedings, Biopharmaceutical Section, 2247 - 2269. https://ww2.amstat.org/MembersOnly/proceedings/2016/data/assets/pdf/389675.pdf”
