+++ paper/paper.bib
@@ -163,9 +163,9 @@ doi = {10.1007/978-981-10-7826-2_11}
 }
 
 @article{FKL2016,
-author = {Fries, M and Kracht, K and Li, J},
-title = {Safety monitoring methodology in the premarketing setting},
-journal = {Proceedings of JSM},
+author = {Fries, M and Kracht, K and Li, J and Munsaka, M S and Sanchez-Kam, M and Singh, K P and Wang, W and Whalen, E and Zhou, K. },
+title = {Safety monitoring methodology in the Premarketing Setting},
+journal = {JSM Proceedings, Biopharmaceutical Section},
 pages = {2247-2269},
 year  = {2016},
 publisher = {Google Scholar},
