### 2020-11-25 rcarragh/c212#8 @rcarragh
I've updated the reference to include the complete author list.
I'm not sure why I had an incomplete list but apologies to those left off the original.
