### 2020-11-25 rcarragh/c212#7 @rcarragh
I've added in a couple of paragraphs in the Summary section (paragraphs 3, 4 and the start of 5) about medical dictionaries, leading up to the methods.
