### 2020-11-25 #2706 @rcarragh
@MelvinSMunsaka, @csoneson - I have put some responses into the issues I created for the comments (rcarragh/c212#7, rcarragh/c212#8, rcarragh/c212#9)

Briefly the following changes have been made to the paper:
1) I have added a couple of paragraphs in the Summary section explaining medical dictionaries and the motivation for the methods
2) I've fixed the reference author issue
3) I've added a section regarding performance and memory requirements in the Software Details and Availability section to give some guidance.
