### 2020-12-01 rcarragh/c212#9 @MelvinSMunsaka
@rcarragh I have reviewed the response and suggested future resolve to my comment and it acceptable.
