### 2020-12-01 rcarragh/c212#7 @MelvinSMunsaka
@rcarragh I have reviewed the revised text and it is acceptable. You can also close this comment.
