### 2020-12-01 #2706 @csoneson
Thanks @MelvinSMunsaka! It seems that you agreed with @rcarragh's changes and that the issues created in response to your comments have been closed, and all the boxes in the checklist are checked. Just to make sure I'm not missing anything, could you just confirm whether you are happy with the state of the submission now, or whether there are things that you would still like to see addressed before it should be accepted? Thanks!
