### 2020-12-02 #2706 @csoneson
👋 @rcarragh - I've made a PR [here](https://github.com/rcarragh/c212/pull/10) with a couple of small fixes to the bibliography, could you please check and merge?
