### 2020-12-02 #2706 @csoneson
The next steps will then be:
- [ ] Check that the Zenodo archive has the correct metadata: the title and author list should match those of the paper; you may also add the authors' ORCID.
