### 2020-12-03 #2706 @rcarragh
@csoneson - I've done the following:

Tagged release of software in github: 1.00
I have archived the corresponding software on Zenodo
The title and author on Zenodo match the paper; Co-author does not have an ORCID.
Zenodo archive DOI: 10.5281/zenodo.4304831

I hope this is correct, thanks.
