### 2020-12-04 #2706 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
