### 2020-12-04 #2706 @rcarragh
I checked the proof and crossref in openjournals/joss-papers#1958 and they look good from my point of view.
